+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [3.67.20-6]
+
+## Fixed
+
+- issue with multiple optional entities
+- better fuzzy comparison
+- chatgpt issue
+
 ## [3.67.19]
 
 ### Fixed